// frontend/src/models/style/ElementStyles.ts
export interface IStaticElementStyle {
    backgroundColor: string
}

export interface IClickableElementStyle {
    backgroundColor: string
    hoverColor: string
    pressColor: string
}

export interface IBorderElementStyle {
    borderRadius: number
    thickness: number
    color: string
}

export interface ITextStyle {
    bold: boolean
    color: string
    size: number
}

export interface IIconStyle {
    iconColor: string
    size: number
}